Simplified memory management for volatile parsing part.
There are 2 parts in an ANTLR genrated parser where memory is allocated: the actual parsing (with or w/o creating a parse tree) and the prediction part (via DFA/ATN etc.). The first part is highly volatile as it recreates parse tree instances (the class) on each parser run. In fact also lexer tokens belong to that part, but are already managed via unique pointers. This first part works without any smart pointer now. Instead there is a simple tracker class which holds all created references and frees them when the parser is reset or destroyed. This is a bit less optimal if the parser is set to create no parse tree, as created rule context objects are not freed immediately (like with smart pointers), but during reset. On the other hand this change gives (depending on the input) a nice speed up (0%-100%, after the warm up phase). Additionally memory consumption drops by a good amount.

Everything in the simulartors (and interpreters) remains unchanged. This is the shared prediction part.

+++ runtime/Cpp/runtime/src/atn/LexerATNSimulator.cpp
@@ -74,7 +74,7 @@ int LexerATNSimulator::match_calls = 0;
 
 LexerATNSimulator::LexerATNSimulator(const ATN &atn, std::vector<dfa::DFA> &decisionToDFA,
                                      PredictionContextCache &sharedContextCache)
-: LexerATNSimulator(nullptr, atn, decisionToDFA, sharedContextCache) {
+  : LexerATNSimulator(nullptr, atn, decisionToDFA, sharedContextCache) {
 }
 
 LexerATNSimulator::LexerATNSimulator(Lexer *recog, const ATN &atn, std::vector<dfa::DFA> &decisionToDFA,
@@ -120,9 +120,9 @@ void LexerATNSimulator::reset() {
 }
 
 void LexerATNSimulator::clearDFA() {
-  int size = (int)_decisionToDFA.size();
+  size_t size = _decisionToDFA.size();
   _decisionToDFA.clear();
-  for (int d = 0; d < size; ++d) {
+  for (size_t d = 0; d < size; ++d) {
     _decisionToDFA.push_back(dfa::DFA(atn.getDecisionState(d), d));
   }
 }
